requirements.txt
requests
pandas
openpyxl
python-dotenv
pathlib